Write the version when upgrading
