+++ src/key.rs
@@ -33,8 +33,8 @@ impl Key {
         Self::new(index, NodeId::metadata())
     }
 
-    pub const fn version(index: u16) -> Self {
-        Self::new(index, NodeId::version())
+    pub const fn version() -> Self {
+        Self::new(u16::MAX, NodeId::version())
     }
 
     pub const fn updated(index: u16, item: u32) -> Self {
